Sandbox.Transactions (Teller Sandbox API v0.1.0)

Context for transactions

Functions

filter_transactions(conn, id, from_id, count)

@spec filter_transactions(map(), binary(), number() | nil, binary() | nil) ::
  {:ok,
   [
     %Sandbox.Transactions.Transaction{
       account_id: term(),
       amount: term(),
       date: term(),
       description: term(),
       details: term(),
       id: term(),
       links: term(),
       running_balance: term(),
       status: term(),
       type: term()
     }
   ]}
  | :error
  | {:error, binary()}

get_balance(conn, id)

@spec get_balance(map(), binary()) ::
  {:ok,
   %Sandbox.Transactions.Balances{
     account_id: term(),
     available: term(),
     ledger: term(),
     links: term()
   }}
  | :error

Gets accounts balance

get_transaction(conn, account_id, transaction_id)

@spec get_transaction(Plug.Conn.t(), binary(), binary()) ::
  :error
  | {:ok,
     %Sandbox.Transactions.Transaction{
       account_id: term(),
       amount: term(),
       date: term(),
       description: term(),
       details: term(),
       id: term(),
       links: term(),
       running_balance: term(),
       status: term(),
       type: term()
     }}

Returns a transaciont by ID

list_transactions(conn, id)

@spec list_transactions(map(), binary()) ::
  {:ok,
   [
     %Sandbox.Transactions.Transaction{
       account_id: term(),
       amount: term(),
       date: term(),
       description: term(),
       details: term(),
       id: term(),
       links: term(),
       running_balance: term(),
       status: term(),
       type: term()
     }
   ]}
  | :error

List all transactions. Transactions are calculated from the end of the yeary to 90 days before today.

So transactions are consistent within a year, it starts calculating transactions from a start balance and adds from 0 to 5 transactions per day.

Transactions are calculated as an Integer and divided by 100 so we don't have the problem of Float calculations.

Every transactions is calculated using :erlang.phash2(), which hashses any string and returns an Integer.

Transactions could be filtered using using from_id and count from id will show the next transaction after that id and it will return the number of transactions by count.